no message
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk@496 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ ac/source/src/clientgame.cpp
@@ -701,7 +701,7 @@ void showmastermenu(int m) // 0=kick, 1=ban
 
 	loopv(players)
 	{
-		if(players[i]) menumanual(menu, i, players[i]->name, (char *) i); // ugly hack
+		if(players[i]) menumanual(menu, i, players[i]->name, (char *)(size_t)i); // ugly hack
 	};
 	menuset(menu);
 };
